Update README.md
This Auto Insurance project is the web- based application software programmed using Java Technology. It manages the activities related to insurance of the people. All the information and data catch up which providing any insurance policy to any company, any human being or any other product such as vehicles etc can be saved in to the database for which oracle is used as a backend. Later on, the insurance holder for any product or insurance holder customers can view the details also using their login details. For rest of the users, the information of each other will keep confidential.